Finished basic handling of items in gear list
Fleshed out the getItems() in CharacterBuilder, and finished the simple
output of all items, without any specific handling.
Minor edits to GearItem and Head
Created a NullGearItem class that allows for the slot to be filled with
a usable object, without limited functionality. Could be useful object
to check for audits and is being used to handle output like any other
item.